Breakout code into more files, fix tests, remove more unneeded code
